FA-37 - refactor(RecoverySeedBoxPage): Organzie the files
