+++ lib/auth/onboard_create_account_page.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:ui_library/ui_library_export.dart';
-import 'package:uplink/auth/onboard_recovery_seed_page.dart';
+import 'package:uplink/auth/onboard_recovery_seed_page/onboard_recovery_seed_page.dart';
 
 class OnboardCreateAccountPage extends StatefulWidget {
   const OnboardCreateAccountPage({Key? key}) : super(key: key);
